feat(windows_foundation): add `getWindowsBuildNumber` function
