+++ packages/windows_ai/test/collections_test.dart
@@ -13,8 +13,6 @@ import 'package:windows_ai/windows_ai.dart';
 import 'package:windows_foundation/internal.dart';
 import 'package:windows_foundation/windows_foundation.dart';
 
-import 'helpers.dart';
-
 void main() {
   if (!isWindowsRuntimeAvailable()) {
     print('Skipping tests because Windows Runtime is not available.');
