+++ packages/windows_foundation/pubspec.yaml
@@ -20,8 +20,9 @@ platforms:
   windows:
 
 dependencies:
-  ffi: ^2.0.2
-  win32: ^5.0.4
+  ffi: ^2.1.0
+  win32: ^5.0.6
+  win32_registry: ^1.1.1
   windows_data: ^0.2.0
   windows_devices: ^0.2.0
   windows_graphics: ^0.2.0
